chatbot-ai_agent/backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.10
python-dotenv==1.1.1
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic-ai-slim[groq]
